// frontend/src/client/models/CodingAssignRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CodingAssignRequest = {
    project: string;
    fragment_id: string;
    codigo: string;
    cita: string;
    fuente?: (string | null);
    memo?: (string | null);
};