// frontend/src/client/models/IngestRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IngestRequest = {
    project: string;
    inputs: Array<string>;
    meta_json?: (string | null);
    batch_size?: number;
    min_chars?: number;
    max_chars?: number;
    run_id?: (string | null);
};